fixes #62 64-bit windows heap corruption

+++ sspi_windows.go
@@ -88,8 +88,8 @@ type TimeStamp struct {
 }
 
 type SecHandle struct {
-	dwLower uint32
-	dwUpper uint32
+	dwLower uintptr
+	dwUpper uintptr
 }
 
 type SecBuffer struct {
